Add anchor tags to Header plus tweak style, add basket file  and define variables, along with useState
