+++ src/App.css
@@ -6,6 +6,8 @@
  }
 
 
+
+
 html{
   background-color: azure;
 }
@@ -14,6 +16,8 @@ html{
   justify-content: space-around;
   background-color: #ffe8d6;
 }
+
+
 
 .links {
   display: flex;
@@ -26,6 +30,16 @@ html{
   font-family: 'Bubblegum Sans', cursive;
  color: #cb997e;
  font-size: 25px;
+}
+
+.links a {
+  text-decoration: none;
+  color: #cb997e;
+}
+
+.links a:hover {
+  text-decoration: none;
+  color: #b1907f;
 }
 
 #title {
@@ -64,3 +78,13 @@ html{
   width: 50%;
   margin: 0 auto;
 }
+
+.btn-check:focus+.btn-primary, 
+.btn-primary:focus {
+  color: #fff;
+  background-color: #b7b7a4 !important;
+  border-color: #6B705C !important;
+  /* box-shadow: 0 0 0 0.25rem rgb(49 132 253 / 50%); */
+}
+
+
